refactor: Shopify 연결 검증 계약과 클라이언트 구성을 정리한다 - ShopifyPingClient.ping ShopifyConnectionVerifier 구현체로 정리하고 ObjectMapper 주입 및 Javadoc 설명을 보강함 - ShopifyOrderClient.countOrdersSince 와 ShopifyFulfillmentApiClient.createFulfillment 직렬화 구성을 주입 방식으로 통일하고 helper 메서드 설명을 추가함 - ShopifyOrderSyncService.syncOrders 와 ShopifyFulfillmentSender.sendBulk 흐름이 정리된 계약을 사용하도록 맞추고 관련 테스트를 갱신함 - IntegrationTest flow 와 ShopifyPingClientTest.ping 등 회귀 테스트를 갱신하고 전체 테스트 통과를 확인함
