+++ src/main/java/com/conk/integration/query/service/ShopifyPingClient.java
@@ -1,6 +1,7 @@
 package com.conk.integration.query.service;
 
 import com.conk.integration.command.infrastructure.config.ShopifyProperties;
+import com.conk.integration.common.channel.ShopifyConnectionVerifier;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
@@ -13,17 +14,21 @@ import org.springframework.web.client.RestTemplate;
 
 import java.util.Map;
 
-// Shopify GraphQL API에 경량 ping을 보내 자격증명 유효성을 확인하는 query 전용 클라이언트다.
+/**
+ * Shopify GraphQL API에 경량 ping을 보내 자격증명 유효성을 확인하는 query 전용 클라이언트다.
+ */
 @Service
 @RequiredArgsConstructor
-public class ShopifyPingClient {
+public class ShopifyPingClient implements ShopifyConnectionVerifier {
 
-    // shop.id만 조회하는 최소 쿼리로 네트워크/서버 부담을 최소화한다.
+    /**
+     * shop.id만 조회하는 최소 쿼리로 네트워크/서버 부담을 최소화한다.
+     */
     private static final String PING_QUERY = "{ shop { id } }";
 
     private final RestTemplate restTemplate;
     private final ShopifyProperties properties;
-    private final ObjectMapper objectMapper = new ObjectMapper();
+    private final ObjectMapper objectMapper;
 
     /**
      * Shopify GraphQL API에 경량 ping을 보내 자격증명 유효성을 확인한다.
@@ -32,6 +37,7 @@ public class ShopifyPingClient {
      * @param accessToken Shopify Admin API 액세스 토큰
      * @return 연결 성공 시 true, HTTP 오류/네트워크 실패 등 모든 예외 시 false
      */
+    @Override
     public boolean ping(String storeName, String accessToken) {
         try {
             String body = objectMapper.writeValueAsString(Map.of("query", PING_QUERY));
@@ -49,6 +55,12 @@ public class ShopifyPingClient {
         }
     }
 
+    /**
+     * Shopify ping 요청에 필요한 인증/콘텐츠 타입 헤더를 생성한다.
+     *
+     * @param accessToken Shopify Admin API 액세스 토큰
+     * @return Shopify 요청 헤더
+     */
     private HttpHeaders buildHeaders(String accessToken) {
         HttpHeaders headers = new HttpHeaders();
         headers.set("X-Shopify-Access-Token", accessToken);
