+++ src/test/java/com/conk/integration/command/application/service/shopify/ShopifyFulfillmentSenderTest.java
@@ -6,8 +6,8 @@ import com.conk.integration.command.domain.aggregate.EasypostShipmentInvoice;
 import com.conk.integration.command.domain.aggregate.enums.OrderChannel;
 import com.conk.integration.command.domain.aggregate.enums.CarrierType;
 import com.conk.integration.command.infrastructure.service.ShopifyFulfillmentApiClient;
-import com.conk.integration.query.dto.FulfillmentTargetDto;
-import com.conk.integration.query.dto.ShopifyCredentialDto;
+import com.conk.integration.command.application.dto.FulfillmentTargetDto;
+import com.conk.integration.common.channel.dto.ShopifyCredentialDto;
 import com.conk.integration.query.service.ChannelApiQueryService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
